Begin support of bpToPx for translocation features
